Refactor: Move code to lineorg client from handlers
