+++ src/backend/api/Fusion.Resources.Api/FusionEvents/LineOrgOrgUnitHandler.cs
@@ -3,6 +3,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Fusion.Resources.Domain;
 using Microsoft.Extensions.Caching.Memory;
+using Fusion.Resources.Application.LineOrg;
 
 namespace Fusion.Resources.Api
 {
@@ -23,7 +24,7 @@ namespace Fusion.Resources.Api
         public async Task ProcessMessageAsync(MessageContext ctx, string? body, CancellationToken cancellationToken)
         {
             await orgUnitCache.ClearOrgUnitCacheAsync();
-            cache.Remove(Domain.GetDepartments.Handler.OrgUnitsMemCacheKey);
+            cache.Remove(LineOrgClient.OrgUnitsMemCacheKey);
         }
     }
 }
