+++ src/backend/api/Fusion.Resources.Domain/Extensions/CompareUtils.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Fusion.Resources
+{
+    public static class CompareUtils
+    {
+        public static bool ContainsNullSafe(this string? value, string? comp)
+        {
+            return (value ?? "").Contains(comp ?? "", StringComparison.OrdinalIgnoreCase);
+        }
+    }
+}
